Refactored GLFW callbacks
